requirements.txt
Flask>=3.0.0
opencv-python>=4.8.0
deepface>=0.0.90
tensorflow>=2.15.0
numpy>=1.24.0
Pillow>=10.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
matplotlib>=3.7.0
pandas>=2.0.0
scikit-learn>=1.3.0
imutils>=0.5.4
requests>=2.30.0
python-dateutil>=2.8.0
